+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 1.2
+
+Adds two new commands for viewing:
+
+- Working tree changes for the current project root
+- File history for the active tab's file
+
+The working tree changes command will only work if you have set the `difftool` for you `git` installation. It could also open another app if it is not configured to use Kaleidoscope.
+
 ## Version 1.1
 
 Adds notifications when adding or clearing files and when launching Kaleidoscope.